assignment/src/components/api-spec/api-spec.spec.ts
import { spy, assert } from 'sinon';
import { expect } from 'chai';
import Component from 'vue-class-component';
import { ComponentTest, MockLogger } from '../../util/component-test';
import { ApiSpecComponent } from './api-spec';

let loggerSpy = spy();

@Component({
  template: require('./api-spec.html')
})
class MockAboutComponent extends ApiSpecComponent {
  constructor() {
    super();
    this.logger = new MockLogger(loggerSpy);
  }
}

describe('About component', () => {
  let directiveTest: ComponentTest;

  beforeEach(() => {
    directiveTest = new ComponentTest('<div><api-spec></api-spec></div>', { 'about': MockAboutComponent });
  });

  it('should render correct contents', async () => {
    debugger;
    directiveTest.createComponent();

    await directiveTest.execute((vm) => {
      expect(vm.$el.querySelector('.repo-link').getAttribute('href')).to.equal('https://github.com/ducksoupdev/vue-webpack-typescript');
      assert.calledWith(loggerSpy, 'api-spec is ready!');
    });
  });
});
